Refactor: importation consistency for `partition_pdf()` and `partition_image()` (#2282)
Closes #2278. This PR also removes the `extract_tables_in_pdf` mentioned
in issue #2280.
