+++ CHANGELOG.md
@@ -1,10 +1,12 @@
-## 0.11.5-dev0
+## 0.11.5-dev1
 
 ### Enhancements
 
 ### Features
 
 ### Fixes
+
+* **Fix `partition_pdf()` and `partition_image()` importation issue.** Reorganize `pdf.py` and `image.py` modules to be consistent with other types of document import code.
 
 ## 0.11.4
