+++ test_unstructured/partition/pdf_image/test_chipper.py
@@ -1,6 +1,6 @@
 import pytest
 
-from unstructured.partition.pdf_image import pdf
+from unstructured.partition import pdf
 from unstructured.partition.utils.constants import PartitionStrategy
