+++ test_unstructured/partition/pdf_image/test_image.py
@@ -10,7 +10,8 @@ from unstructured_inference.inference import layout
 from test_unstructured.unit_utils import assert_round_trips_through_JSON, example_doc_path
 from unstructured.chunking.title import chunk_by_title
 from unstructured.documents.elements import ElementType
-from unstructured.partition.pdf_image import image, ocr, pdf
+from unstructured.partition import image, pdf
+from unstructured.partition.pdf_image import ocr
 from unstructured.partition.utils.constants import (
     UNSTRUCTURED_INCLUDE_DEBUG_METADATA,
     PartitionStrategy,
@@ -303,7 +304,7 @@ def test_partition_image_metadata_date(
 ):
     mocked_last_modification_date = "2029-07-05T09:24:28"
     mocker.patch(
-        "unstructured.partition.pdf_image.pdf.get_last_modified_date",
+        "unstructured.partition.pdf.get_last_modified_date",
         return_value=mocked_last_modification_date,
     )
     elements = image.partition_image(filename=filename)
@@ -317,7 +318,7 @@ def test_partition_image_with_hi_res_strategy_metadata_date(
 ):
     mocked_last_modification_date = "2029-07-05T09:24:28"
     mocker.patch(
-        "unstructured.partition.pdf_image.pdf.get_last_modified_date",
+        "unstructured.partition.pdf.get_last_modified_date",
         return_value=mocked_last_modification_date,
     )
     elements = image.partition_image(filename=filename, stratefy=PartitionStrategy.HI_RES)
@@ -333,7 +334,7 @@ def test_partition_image_metadata_date_custom_metadata_date(
     expected_last_modification_date = "2009-07-05T09:24:28"
 
     mocker.patch(
-        "unstructured.partition.pdf_image.pdf.get_last_modified_date",
+        "unstructured.partition.pdf.get_last_modified_date",
         return_value=mocked_last_modification_date,
     )
     elements = image.partition_image(
@@ -352,7 +353,7 @@ def test_partition_image_with_hi_res_strategy_metadata_date_custom_metadata_date
     expected_last_modification_date = "2009-07-05T09:24:28"
 
     mocker.patch(
-        "unstructured.partition.pdf_image.pdf.get_last_modified_date",
+        "unstructured.partition.pdf.get_last_modified_date",
         return_value=mocked_last_modification_date,
     )
     elements = image.partition_image(
@@ -370,7 +371,7 @@ def test_partition_image_from_file_metadata_date(
 ):
     mocked_last_modification_date = "2029-07-05T09:24:28"
     mocker.patch(
-        "unstructured.partition.pdf_image.pdf.get_last_modified_date_from_file",
+        "unstructured.partition.pdf.get_last_modified_date_from_file",
         return_value=mocked_last_modification_date,
     )
     with open(filename, "rb") as f:
@@ -385,7 +386,7 @@ def test_partition_image_from_file_with_hi_res_strategy_metadata_date(
 ):
     mocked_last_modification_date = "2029-07-05T09:24:28"
     mocker.patch(
-        "unstructured.partition.pdf_image.pdf.get_last_modified_date_from_file",
+        "unstructured.partition.pdf.get_last_modified_date_from_file",
         return_value=mocked_last_modification_date,
     )
 
@@ -403,7 +404,7 @@ def test_partition_image_from_file_metadata_date_custom_metadata_date(
     expected_last_modification_date = "2009-07-05T09:24:28"
 
     mocker.patch(
-        "unstructured.partition.pdf_image.pdf.get_last_modified_date_from_file",
+        "unstructured.partition.pdf.get_last_modified_date_from_file",
         return_value=mocked_last_modification_date,
     )
     with open(filename, "rb") as f:
@@ -423,7 +424,7 @@ def test_partition_image_from_file_with_hi_res_strategy_metadata_date_custom_met
     expected_last_modification_date = "2009-07-05T09:24:28"
 
     mocker.patch(
-        "unstructured.partition.pdf_image.pdf.get_last_modified_date_from_file",
+        "unstructured.partition.pdf.get_last_modified_date_from_file",
         return_value=mocked_last_modification_date,
     )
     with open(filename, "rb") as f:
