+++ test_unstructured/partition/pdf_image/test_pdf.py
@@ -20,9 +20,9 @@ from unstructured.documents.elements import (
     Text,
     Title,
 )
-from unstructured.partition import strategies
-from unstructured.partition.pdf_image import ocr, pdf, pdfminer_processing
-from unstructured.partition.pdf_image.pdf import get_uris_from_annots
+from unstructured.partition import pdf, strategies
+from unstructured.partition.pdf import get_uris_from_annots
+from unstructured.partition.pdf_image import ocr, pdfminer_processing
 from unstructured.partition.utils.constants import (
     UNSTRUCTURED_INCLUDE_DEBUG_METADATA,
     PartitionStrategy,
@@ -679,7 +679,7 @@ def test_partition_pdf_metadata_date(
     )
 
     mocker.patch(
-        "unstructured.partition.pdf_image.pdf.get_the_last_modification_date_pdf_or_img",
+        "unstructured.partition.pdf.get_the_last_modification_date_pdf_or_img",
         return_value=mocked_last_modification_date,
     )
