+++ test_unstructured/partition/test_auto.py
@@ -693,10 +693,7 @@ def test_file_specific_produces_correct_filetype(filetype: FileType):
         extension if filetype not in FILETYPE_TO_MODULE else FILETYPE_TO_MODULE[filetype]
     )
     fun_name = "partition_" + filetype_module
-    if filetype_module in ["pdf", "image"]:
-        module = import_module(f"unstructured.partition.pdf_image.{filetype_module}")  # noqa
-    else:
-        module = import_module(f"unstructured.partition.{filetype_module}")  # noqa
+    module = import_module(f"unstructured.partition.{filetype_module}")  # noqa
     fun = eval(f"module.{fun_name}")
     for file in pathlib.Path("example-docs").iterdir():
         if file.is_file() and file.suffix == f".{extension}":
