+++ unstructured/nlp/partition.py
@@ -1,5 +1,5 @@
 # flake8: noqa
-from unstructured.partition.pdf_image.pdf import partition_pdf  # noqa
+from unstructured.partition.pdf import partition_pdf  # noqa
 from unstructured.partition.text_type import (  # noqa
     is_bulleted_text,
     is_possible_narrative_text,
