+++ unstructured/partition/pdf.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import contextlib
 import io
 import os
@@ -137,7 +139,6 @@ def partition_pdf(
     extract_images_in_pdf: bool = False,
     extract_element_types: Optional[List[str]] = None,
     image_output_dir_path: Optional[str] = None,
-    hi_res_model_name: Optional[str] = None,
     **kwargs,
 ) -> List[Element]:
     """Parses a pdf document into a list of interpreted elements.
@@ -170,14 +171,9 @@ def partition_pdf(
     extract_images_in_pdf
         If True and strategy=hi_res, any detected images will be saved in the path specified by
         image_output_dir_path.
-    extract_tables_in_pdf
-        If True and strategy=hi_res, any detected tables will be saved in the path specified by
-        image_output_dir_path.
     image_output_dir_path
-        If extract_images_in_pdf=True (or extract_tables_in_pdf=True) and strategy=hi_res, any
-        detected images or tables will be saved in the given path
-    hi_res_model_name
-        The layout detection model used when partitioning strategy is set to `hi_res`.
+        If extract_images_in_pdf=True and strategy=hi_res, any detected images or tables
+        will be saved in the given path
     """
 
     exactly_one(filename=filename, file=file)
@@ -195,7 +191,6 @@ def partition_pdf(
         extract_images_in_pdf=extract_images_in_pdf,
         extract_element_types=extract_element_types,
         image_output_dir_path=image_output_dir_path,
-        hi_res_model_name=hi_res_model_name,
         **kwargs,
     )
 
@@ -241,14 +236,13 @@ def _partition_pdf_or_image_local(
     include_page_breaks: bool = False,
     languages: Optional[List[str]] = None,
     ocr_mode: str = OCRMode.FULL_PAGE.value,
-    model_name: Optional[str] = None,  # to be deprecated in favor of `hi_res_model_name`
+    model_name: Optional[str] = None,
     metadata_last_modified: Optional[str] = None,
     pdf_text_extractable: bool = False,
     extract_images_in_pdf: bool = False,
     extract_element_types: Optional[List[str]] = None,
     image_output_dir_path: Optional[str] = None,
     pdf_image_dpi: Optional[int] = None,
-    hi_res_model_name: Optional[str] = None,
     **kwargs,
 ) -> List[Element]:
     """Partition using package installed locally"""
@@ -271,12 +265,10 @@ def _partition_pdf_or_image_local(
 
     ocr_languages = prepare_languages_for_tesseract(languages)
 
-    hi_res_model_name = (
-        hi_res_model_name or model_name or default_hi_res_model(infer_table_structure)
-    )
+    model_name = model_name or default_hi_res_model(infer_table_structure)
     if pdf_image_dpi is None:
-        pdf_image_dpi = 300 if hi_res_model_name == "chipper" else 200
-    if (pdf_image_dpi < 300) and (hi_res_model_name == "chipper"):
+        pdf_image_dpi = 300 if model_name == "chipper" else 200
+    if (pdf_image_dpi < 300) and (model_name == "chipper"):
         logger.warning(
             "The Chipper model performs better when images are rendered with DPI >= 300 "
             f"(currently {pdf_image_dpi}).",
@@ -286,7 +278,7 @@ def _partition_pdf_or_image_local(
         inferred_document_layout = process_file_with_model(
             filename,
             is_image=is_image,
-            model_name=hi_res_model_name,
+            model_name=model_name,
             pdf_image_dpi=pdf_image_dpi,
         )
 
@@ -299,7 +291,7 @@ def _partition_pdf_or_image_local(
         else:
             merged_document_layout = inferred_document_layout
 
-        if hi_res_model_name.startswith("chipper"):
+        if model_name.startswith("chipper"):
             # NOTE(alan): We shouldn't do OCR with chipper
             final_document_layout = merged_document_layout
         else:
@@ -316,7 +308,7 @@ def _partition_pdf_or_image_local(
         inferred_document_layout = process_data_with_model(
             file,
             is_image=is_image,
-            model_name=hi_res_model_name,
+            model_name=model_name,
             pdf_image_dpi=pdf_image_dpi,
         )
         if hasattr(file, "seek"):
@@ -330,7 +322,7 @@ def _partition_pdf_or_image_local(
         else:
             merged_document_layout = inferred_document_layout
 
-        if hi_res_model_name.startswith("chipper"):
+        if model_name.startswith("chipper"):
             # NOTE(alan): We shouldn't do OCR with chipper
             final_document_layout = merged_document_layout
         else:
@@ -347,7 +339,7 @@ def _partition_pdf_or_image_local(
             )
 
     # NOTE(alan): starting with v2, chipper sorts the elements itself.
-    if hi_res_model_name == "chipper":
+    if model_name == "chipper":
         kwargs["sort_mode"] = SORT_MODE_DONT
 
     final_document_layout = clean_pdfminer_inner_elements(final_document_layout)
@@ -417,7 +409,7 @@ def _partition_pdf_or_image_local(
             ).strip()
             # NOTE(alan): with chipper there are parent elements with no text we don't want to
             # filter those out and leave the children orphaned.
-            if el.text or isinstance(el, PageBreak) or hi_res_model_name.startswith("chipper"):
+            if el.text or isinstance(el, PageBreak) or model_name.startswith("chipper"):
                 out_elements.append(cast(Element, el))
 
     return out_elements
@@ -436,7 +428,6 @@ def partition_pdf_or_image(
     extract_images_in_pdf: bool = False,
     extract_element_types: Optional[List[str]] = None,
     image_output_dir_path: Optional[str] = None,
-    hi_res_model_name: Optional[str] = None,
     **kwargs,
 ) -> List[Element]:
     """Parses a pdf or image document into a list of interpreted elements."""
@@ -498,7 +489,6 @@ def partition_pdf_or_image(
                 extract_images_in_pdf=extract_images_in_pdf,
                 extract_element_types=extract_element_types,
                 image_output_dir_path=image_output_dir_path,
-                hi_res_model_name=hi_res_model_name,
                 **kwargs,
             )
             out_elements = _process_uncategorized_text_elements(elements)
